% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PV.R
\name{drawPVs}
\alias{drawPVs}
\alias{drawPVs.summary.mmlMeans}
\alias{drawPVs.mmlMeans}
\alias{drawPVs.summary.mmlCompositeMeans}
\alias{drawPVs.mmlCompositeMeans}
\title{Draw plausible values (PVs) from an mml fit}
\usage{
drawPVs(
  x,
  npv,
  pvVariableNameSuffix = "_dire",
  stochasticBeta = deprecated(),
  ...
)

\method{drawPVs}{summary.mmlMeans}(
  x,
  npv = 5L,
  pvVariableNameSuffix = "_dire",
  stochasticBeta = deprecated(),
  ...
)

\method{drawPVs}{mmlMeans}(
  x,
  npv = 5L,
  pvVariableNameSuffix = "_dire",
  stochasticBeta = deprecated(),
  ...
)

\method{drawPVs}{summary.mmlCompositeMeans}(
  x,
  npv = 20L,
  pvVariableNameSuffix = "_dire",
  stochasticBeta = deprecated(),
  ...
)

\method{drawPVs}{mmlCompositeMeans}(
  x,
  npv = 20L,
  pvVariableNameSuffix = "_dire",
  stochasticBeta = deprecated(),
  ...
)
}
\arguments{
\item{x}{a fit from a call to \code{\link{mml}}}

\item{npv}{integer indicating the number of plausible values to draw}

\item{pvVariableNameSuffix}{suffix added to new PV variables after construct name and before the plausible value ID. For example, if there is a construct \code{math} and the suffix is the default \code{_dire}, then the fourth plausible value would have a column name, \code{math_dire4}.}

\item{stochasticBeta}{deprecated.}

\item{\dots}{additional parameters passed on to methods.}
}
\value{
the \code{stuDat} attribute of the \code{\link{mml}} fit object, \code{x}, with columns for the new plausible values.
}
\description{
Draw plausible values (PVs) from an mml fit
}
\details{
When the argument passed to \code{stocasticBeta} is a data frame then each column is an element that will be used as a
regression coefficient for that index of the coefficients vector. The row index used for the nth PV will be the nth row.
}
\examples{
# See Examples in mml
}
\author{
Paul Bailey, Sun-joo Lee, and Eric Buehler
}
